# services/puzzle-generator/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Chess puzzle generator service (Lichess, AI via OpenRouter, and Hugging Face dataset).",
        "title": "Puzzle Generator API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
        "/health": {
            "get": {
                "description": "Health probe endpoint",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "health"
                ],
                "summary": "Service health",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.HealthResponse"
                        }
                    }
                }
            }
        },
        "/puzzle": {
            "get": {
                "description": "Returns a random puzzle (Lichess source) filtered by difficulty",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "puzzle"
                ],
                "summary": "Get puzzle by difficulty",
                "parameters": [
                    {
                        "enum": [
                            "easy",
                            "medium",
                            "hard"
                        ],
                        "type": "string",
                        "description": "easy|medium|hard",
                        "name": "difficulty",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Puzzle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/puzzle/ai": {
            "post": {
                "description": "Generates one custom puzzle from OpenRouter model",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "puzzle"
                ],
                "summary": "Generate puzzle from AI",
                "parameters": [
                    {
                        "description": "AI puzzle request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/models.AIPuzzleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Puzzle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/puzzle/daily": {
            "get": {
                "description": "Returns Lichess daily puzzle",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "puzzle"
                ],
                "summary": "Get daily puzzle",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Puzzle"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/puzzle/dataset": {
            "get": {
                "description": "Returns one random puzzle from Hugging Face Lichess dataset",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "puzzle"
                ],
                "summary": "Get puzzle from dataset",
                "parameters": [
                    {
                        "enum": [
                            "easy",
                            "medium",
                            "hard"
                        ],
                        "type": "string",
                        "description": "easy|medium|hard",
                        "name": "difficulty",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Puzzle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/puzzle/{id}": {
            "get": {
                "description": "Returns a puzzle by Lichess puzzle identifier",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "puzzle"
                ],
                "summary": "Get puzzle by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Puzzle ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Puzzle"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/models.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "handlers.HealthResponse": {
            "type": "object",
            "properties": {
                "service": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "version": {
                    "type": "string"
                }
            }
        },
        "models.AIPuzzleRequest": {
            "type": "object",
            "properties": {
                "difficulty": {
                    "$ref": "#/definitions/models.DifficultyLevel"
                },
                "prompt": {
                    "type": "string"
                }
            }
        },
        "models.DifficultyLevel": {
            "type": "string",
            "enum": [
                "easy",
                "medium",
                "hard"
            ],
            "x-enum-varnames": [
                "DifficultyEasy",
                "DifficultyMedium",
                "DifficultyHard"
            ]
        },
        "models.ErrorResponse": {
            "type": "object",
            "properties": {
                "details": {
                    "type": "string"
                },
                "error": {
                    "type": "string"
                }
            }
        },
        "models.Puzzle": {
            "type": "object",
            "properties": {
                "difficulty": {
                    "$ref": "#/definitions/models.DifficultyLevel"
                },
                "fen": {
                    "type": "string"
                },
                "gameUrl": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "initialPly": {
                    "type": "integer"
                },
                "moves": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "nbPlays": {
                    "type": "integer"
                },
                "popularity": {
                    "type": "integer"
                },
                "rating": {
                    "type": "integer"
                },
                "ratingDeviation": {
                    "type": "integer"
                },
                "source": {
                    "type": "string"
                },
                "themes": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        }
    }
}